Rename namespaces and repositorys-->repositories
